refactor(clawdstrike-policy-event): extract validate from edr/receipt/mod.rs
